story component update 2
also changed "Story" to "Stories" for consistency
