+++ src/components/Layout/Layout.jsx
@@ -8,7 +8,6 @@
 import React from "react"
 import Navbar from "../Navbar"
 import Footer from "../Footer"
-import Stories from "../Stories"
 import SEO from "../SEO"
 import PropTypes from "prop-types"
 import "normalize.css"
@@ -21,7 +20,6 @@ const Layout = ({ children, seo, slug }) => {
       <Navbar slug={slug} />
       <SEO title={siteTitle} description={siteDescription} />
       {children}
-      <Stories />
       <Footer />
     </>
   )
